Adding projects from Computer Architecture Class
In this project, I have uploaded two coding projects from my computer architecture class. For the first project (covidtracker.c), I was asked to write a C program that would take a list of names and create a binary search tree that illustrated who infected whom with COVID-19 based on criteria that was provided. The other project I uploaded is called cachesim.c. In this project, I was given the choice to simulate a cache using either C or Java (doing it in C was rewarded 10 extra credit points due to the added difficulty).