Uploading files from Computer Architecture
The project I uploaded in this commit is written in MIPS assembly code. The project prompt is the same as the covidtracker project I had to complete in C. However, we were asked to complete it in MIPS this time. Although MIPS is not usually used in general software engineering roles, this project shows problem solving skills as it was a significant challenge to read a file and generate a binary search tree with assembly code.